[CALCITE-XXXX] Encapsulate handler generation
Encapsulating MetadataHandler creation in org.apache.calcite.rel.metadta.janino.
Cleaning up method names and java docs.
